Make byte offset for file tests a uin32_t instead of uint64_t

+++ src/descriptors.cpp
@@ -385,7 +385,7 @@ bool FileTestBytes::do_matches(const FileReaderDescriptor &desc) const
 
 
 std::vector<char> FileTestBytes::read_bytes(const std::string &filename,
-	const uint64_t &offset, const uint32_t &length) const
+	const uint32_t &offset, const uint32_t &length) const
 {
 	// Read a number of bytes from the start of the file
 
@@ -418,7 +418,6 @@ std::vector<char> FileTestBytes::read_bytes(const std::string &filename,
 	{
 		int64_t total_bytes_read = in.gcount();
 
-		//ARCS_LOG_ERROR << "Failed to read from file: " << f.what();
 		bytes.resize(0);
 
 		throw FileReadException(f.what(), total_bytes_read + 1);
